requirements.txt
# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# MLOps and Experiment Tracking
mlflow>=2.5.0
dvc>=3.0.0
evidently>=0.4.0

# Workflow Orchestration
apache-airflow>=2.6.0

# Monitoring and Dashboard
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
joblib>=1.3.0
pyyaml>=6.0
